Rename file to test_data
